Initial commit
I am working on uploading all of my photos to photos.google.com. This is
a large undertaking. I cannot do it all in one run. So, I need a way to
mark which photos have been uploaded. Use an xattr to do that.

Also, photos.google.com will only upload files from a single directory
at a time. Make a system where we can copy photos into a directory and
then mark them as uploaded. Copying is expensive and we lose track of
the original (which is the one we want to mark). Use links instead.