Initial commit
Given a YYMMDD, iterate through all of the files in the current
directory and set their accessed and modified dates to YYMMDD101010,
incrementing for each file.

The intent is for photos that were scanned in some random year to be
able to set the dates on those scanned files to the actual day that the
photos were taken. And, to increment the dates to show what order they
were taken in. This is useful for uploading photos to photos.google.com
where they organize the photos by date, rather than file name or
anything else.