Warn if MacOs turns the default prompt on
MacOs ships with a default zsh prompt. As far as I can tell, it is not
possible to override this prompt. So, it has to be manually commented
out. But, updates to MacOs rewrite the file and turn the prompt back on.
If the line gets uncommented, signal a warning.