Major rewrite
This fixes some shell script weaknesses (like iterating on find).

This addresses (not fixes) a problem where attempts to acces github fail
with a 'memory allocation failed' error. Previously, that would stop the
whole script. Now it just ignores the error and moves on to the next repo. Not perfect, but definitely better.